+++ src/components/UI/buyNowBtn/buyNowBtn.scss
