+++ src/components/UI/buyNowBtn/buyNowBtn.tsx
