+++ src/components/UI/carousel/carousel.scss
