+++ src/components/UI/dualSlider/dualSlider.scss
