+++ src/components/UI/dualSlider/dualSlider.tsx
