+++ src/pages/cartPage/components/modalWindow/modalWindow.scss
