+++ src/pages/cartPage/components/pagination/pagination.scss
