+++ src/pages/cartPage/components/productCart/productCart.tsx
@@ -1,8 +1,8 @@
 import './productCart.scss';
-import { products } from '../../data/data';
+import { products } from '../../../../data/data';
 import { useEffect, useState } from 'react';
 import { NavLink } from 'react-router-dom';
-import { Carousel } from '../../UI/carousel/carousel';
+import { Carousel } from '../../../../components/UI/carousel/carousel';
 
 interface ProductCartProps {
   cartId: number;
