+++ src/pages/productsPage/components/categoryItem/categoryItem.scss
