+++ src/pages/productsPage/components/categoryItem/categoryItem.tsx
@@ -1,8 +1,8 @@
 import './categoryItem.scss';
-import { IProduct } from '../../types';
+import { IProduct } from '../../../../types';
 import { useSearchParams } from 'react-router-dom';
-import { products } from '../../data/data';
-import { deleteParam } from '../../utility/utility';
+import { products } from '../../../../data/data';
+import { deleteParam } from '../../../../utility/utility';
 import { useState, useEffect, FC } from 'react';
 
 interface ICategoryItemProps {
