+++ src/pages/productsPage/components/filterPanel/filterPanel.scss
