+++ src/pages/productsPage/components/productCard/productCard.scss
