+++ src/pages/productsPage/components/productCard/video-icon.png
