+++ src/pages/productsPage/components/productViewControl/productViewControl.scss
