+++ src/pages/productsPage/components/productViewControl/productViewControl.tsx
